refactor!: remove Tauri/Expo presets — platform code moves to apps
Delete preset/tauri.ts, preset/expo.ts, storage/tauri.ts, storage/expo.ts.
PlatformPreset interface moves to internal preset/common.ts (not exported).
Remove peerDeps: @tauri-apps/plugin-sql, expo-sqlite, kysely-dialect-tauri.
Remove ambient.d.ts kysely-expo declaration.

Add subpath export: ./transport/browser for apps that need BrowserWebSocket.

nodePreset and testPreset remain as SDK defaults.
Apps define their own BootstrapConfig with platform-specific Kysely dialects.
