+++ package.json
@@ -15,14 +15,6 @@
             "types": "./dist/preset/node.d.ts",
             "import": "./dist/preset/node.js"
         },
-        "./preset/tauri": {
-            "types": "./dist/preset/tauri.d.ts",
-            "import": "./dist/preset/tauri.js"
-        },
-        "./preset/expo": {
-            "types": "./dist/preset/expo.d.ts",
-            "import": "./dist/preset/expo.js"
-        },
         "./preset/test": {
             "types": "./dist/preset/test.d.ts",
             "import": "./dist/preset/test.js"
@@ -31,14 +23,6 @@
             "types": "./dist/storage/node.d.ts",
             "import": "./dist/storage/node.js"
         },
-        "./storage/tauri": {
-            "types": "./dist/storage/tauri.d.ts",
-            "import": "./dist/storage/tauri.js"
-        },
-        "./storage/expo": {
-            "types": "./dist/storage/expo.d.ts",
-            "import": "./dist/storage/expo.js"
-        },
         "./storage/sqlite": {
             "types": "./dist/storage/sqlite.d.ts",
             "import": "./dist/storage/sqlite.js"
@@ -46,6 +30,10 @@
         "./storage/schema": {
             "types": "./dist/storage/schema.d.ts",
             "import": "./dist/storage/schema.js"
+        },
+        "./transport/browser": {
+            "types": "./dist/transport/browser.d.ts",
+            "import": "./dist/transport/browser.js"
         },
         "./keystore/node": {
             "types": "./dist/keystore/node.d.ts",
@@ -84,7 +72,6 @@
     },
     "devDependencies": {
         "@changesets/cli": "^2.30.0",
-        "@tauri-apps/plugin-sql": "^2.4.0",
         "@types/better-sqlite3": "^7.6.13",
         "@types/ed2curve": "^0.2.4",
         "@types/node": "^24.12.2",
@@ -97,10 +84,8 @@
         "eslint": "^10.1.0",
         "eslint-config-prettier": "^10.1.8",
         "eslint-plugin-perfectionist": "^5.8.0",
-        "expo-sqlite": "^14.0.6",
         "fast-check": "^4.6.0",
         "husky": "^9.1.7",
-        "kysely-dialect-tauri": "^1.2.1",
         "lint-staged": "^16.3.2",
         "prettier": "^3.2.0",
         "rimraf": "^6.1.3",
@@ -131,25 +116,10 @@
         "zod": "^4.3.6"
     },
     "peerDependencies": {
-        "@tauri-apps/plugin-sql": ">=2.0.0",
-        "better-sqlite3": ">=11.0.0",
-        "expo-sqlite": ">=14.0.0",
-        "kysely-dialect-tauri": ">=1.2.0"
+        "better-sqlite3": ">=11.0.0"
     },
     "peerDependenciesMeta": {
-        "@tauri-apps/plugin-sql": {
-            "optional": true
-        },
         "better-sqlite3": {
-            "optional": true
-        },
-        "expo-sqlite": {
-            "optional": true
-        },
-        "kysely-dialect-tauri": {
-            "optional": true
-        },
-        "kysely-expo": {
             "optional": true
         }
     },
